shop/admin/AI/requirements.txt
Flask==2.2.3
Werkzeug==2.2.3
flask-cors==3.0.10
mysql-connector-python==8.0.33
transformers==4.32.0
scikit-learn==1.3.0
torch==2.0.1
numpy==1.24.4
